Fix #162: Generic tag access now always returns the first value only.
Changed the behavior of the generic tag access methods in ID3v2::Tag;
now they only return the first value in the presence of multiple values.
This is consistent with the other tag types supporting multiple values
(Vorbis, APE).

Since this introduces a semantic change (particularly in the case of
genre()), I am not sure if we should schedule this for the next
major release?