Respect atom type when converting rate tag
TagLib prior to #818 (commit ff28cf276c60)read and wrote the "rate" tag as
text, and switched to reading it as integer value even when the atom class
is a text type. This breaks reading existing files, and can be avoided by
taking the atom type into account.

This fixes issue #885.