Respect MP4::Item type when writing the rate tag
TagLib prior to #818 (commit ff28cf276c60) read and wrote the "rate" tag
as text, and switched to writing the integer value of the MP4::Item.
This breaks writing from applications which supply the value as
StringList, which was the previously implemented API. Applications using
an MP4::Item(UInt) are still supported whithout changes on the application
side.

This is the complimentary writing part for issue #885.