Begin implementing the rest of the File API
This adds:

- ID3v2 versions (v3, v4): working
- Ability to create missing tags (working for ID3v2, not for DIIN)
- Ability to strip tags (incomplete)
- Some tests for these features, and other basic functionality
  (incomplete)

There are still a number of major issues in the code, so I'm pushing
this out to a feature branch for now.  Specifically, the structure of
the code does not closely follow the format specification (as in other
implementations) and is not internally documented, so it is very
difficult to follow.

Specifically chunk type names should follow more closely those in the
spec.  Right now there is only one chunk type (whereas the spec has
several), and the notion / implementation of child chunks seems
incomplete.

Addtionally, before this is merged back to master a number of formatting
changes are needed:

- Spacing as done in the rest of TagLib
- Remove usage of NULL
- Usage of TagLib's containers instead of STL containers
- Use of std::array rather than C arrays where fixed length containers
  are needed
- More complete test coverage

CC #878