Pass separator to both method calls in stringUnion...?!
This doesn't make sense to me, but if e.g. genre().isEmpty() is called
without an argument and then genre(separator) is called, the second call
still gets the default value for separator rather than the passed in
value (verified multiple times across multiple functions in gdb). Seems
like a compiler optimization, but the function is getting invoked both
times, and I *thought* I disabled all optimizations.

Anyway, it's relatively easy to pass the separator argument to each
method call, so do that.